Update_1_MainMenu
A main menu has been implemented. A sample background has been placed, we can change it later. The player can now start the game and will be taken to the first level. (No level exists, so it will send the player to a blank screen). An Option button exists, once the player presses it, they can go back by pressing the "Back" button. The Quit button now exists as well. That is all.