Update_2_PauseMenu
After pressing "Play", you will be taken to Level_0, which should be a blue background with nothing on it. Pressing the "Esc" button will bring up the pause menu. Pressing "Resume" will resume the game, "Menu" will take you to the main menu, and "Quit" will quit the game. Since we can't quit out of Unity, a message saying, "You have quit the game" will be shown in the console. Pressing "Esc" again, while paused, will also resume the game. That is all.