Update 12
Fixed player movement. When going towards house door, it will take you to another scene that should be inside the house. When going towards the wheel, it will take you to scene Level1. I have temporarily added a rope that way when collected, the wheel will check if in inventory will take to new scene (have not implemented yet).